DELFT3D-35990: New directory structure for the binaries: Docker scripts updated

+++ src/engines_gpl/flow2d3d/scripts/run_dflow2d3d_parallel.sh
@@ -16,6 +16,8 @@ function print_usage_info {
     echo "Options:"
     echo "-h, --help"
     echo "       print this help message and exit"
+    echo "--dockerparallel"
+    echo "       A parallel run inside docker"
     echo "<filename>"
     echo "       Delft3D-FLOW configuration filename, default config_d_hydro.xml"
     exit 1
@@ -30,6 +32,7 @@ function print_usage_info {
 ## Defaults
 NPART=0
 configfile=config_d_hydro.xml
+dockerprl=0
 D3D_HOME=
 ulimit -s unlimited
 
@@ -47,6 +50,9 @@ shift
 case $key in
     -h|--help)
     print_usage_info
+    ;;
+    --dockerparallel)
+    dockerprl=1
     ;;
     --D3D_HOME)
     D3D_HOME="$1"
@@ -92,6 +98,7 @@ echo "    Configfile       : $configfile"
 echo "    D3D_HOME         : $D3D_HOME"
 echo "    Working directory: $workdir"
 echo "    nr of parts      : $NPART"
+echo "    Docker parallel  : $dockerprl"
 echo 
 
     #
@@ -104,30 +111,57 @@ libdir=$D3D_HOME/lib
 
 
 export LD_LIBRARY_PATH=$libdir:$LD_LIBRARY_PATH
-export PATH="/opt/mpich2/bin:$bindir:${PATH}"
-export NHOSTS=$NPART
+export PATH="$bindir:${PATH}"
 
-    # Start mpi
-echo " ">$(pwd)/machinefile
-mpd &
-mpdboot -n $NHOSTS
+if [ $dockerprl -eq 1 ]; then
+    #
+    # Parallel in Docker
+    # Assumption: 1 node
+    export PATH=/usr/lib64/mpich/bin:$PATH
+    echo "Starting mpd..."
+    mpd &
+    mpdboot -n $NPART --rsh=/usr/bin/rsh
 
-    # link mpich debug rubbish to /dev/null
-node_number=$NPART
-while test $node_number -ge 1
-do
-   node_number=`expr $node_number - 1`
-   ln -s /dev/null log$node_number.irlog
-done
+    node_number=$NPART
+    while [ $node_number -ge 1 ]; do
+       node_number=`expr $node_number - 1`
+       ln -s /dev/null log$node_number.irlog
+    done
 
-    # Run
     echo "executing:"
     echo "mpirun -np $NPART $bindir/d_hydro $configfile"
-    echo 
-mpirun -np $NPART $bindir/d_hydro $configfile
+          mpirun -np $NPART $bindir/d_hydro $configfile
+else 
+    #
+    # Not in Docker
+    # Start mpi
+    export PATH="/opt/mpich2/bin:${PATH}"
+    export NHOSTS=$NPART
+    echo " ">$(pwd)/machinefile
+    mpd &
+    mpdboot -n $NHOSTS
 
-rm -f log*.irlog
-mpdallexit
+       # link mpich debug rubbish to /dev/null
+    node_number=$NPART
+    while test $node_number -ge 1
+    do
+       node_number=`expr $node_number - 1`
+       ln -s /dev/null log$node_number.irlog
+    done
+
+       # Run
+        echo "executing:"
+        echo "mpirun -np $NPART $bindir/d_hydro $configfile"
+        echo 
+    mpirun -np $NPART $bindir/d_hydro $configfile
+
+    rm -f log*.irlog
+fi
+
+
+if [[ $dockerprl -eq 1 ]]; then
+    mpdallexit
+fi
 
     # Wait until all child processes are finished
 wait
